+++ nnunetv2/training/dataloading/data_loader_3d_random_raters.py
@@ -230,7 +230,7 @@ class nnUNetDataLoader3D_channel_sampler(nnUNetDataLoader):
         if self.transforms is not None:
             with torch.no_grad():
                 with threadpool_limits(limits=1, user_api=None):
-                    data_all = torch.from_numpy(data_all).float()
+                    data_all = torch.from_numpy(data_all)
                     seg_all = torch.from_numpy(seg_all).to(torch.int16)
                     images = []
                     segs = []
